# cmd/shortener-api/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for creating and resolving shortened URLs.",
        "title": "URL Shortener",
        "contact": {
            "name": "Promise",
            "email": "promiseihunna@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8003",
    "basePath": "/",
    "paths": {
        "/api/v1/health": {
            "get": {
                "description": "Returns the health status of the shortener API.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "operationId": "healthCheck",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/links": {
            "get": {
                "description": "Fetch all links record",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Get links",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit pagination records",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "specify pagination page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.GetLinksResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create new shortened URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Shorten URL",
                "parameters": [
                    {
                        "description": "Create Link payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.CreateLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/links/{id}": {
            "get": {
                "description": "Fetch link by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Get link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete link by id",
                "tags": [
                    "links"
                ],
                "summary": "Delete link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "description": "Update link by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Update link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Link id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/{shortCode}": {
            "get": {
                "description": "Redirect the client to the original long URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "links"
                ],
                "summary": "Redirect by short code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short code",
                        "name": "shortCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary redirect to the long URL",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "Destination long URL"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreateLinkRequest": {
            "type": "object",
            "required": [
                "long_url"
            ],
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "long_url": {
                    "type": "string",
                    "maxLength": 20448
                }
            }
        },
        "handler.CreateLinkResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.LinkSample"
                },
                "message": {
                    "type": "string",
                    "example": "Link created successfully!"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.GetLinkResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.LinkSample"
                },
                "message": {
                    "type": "string",
                    "example": "Records fetched successfully!"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.GetLinksResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.LinkSample"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Link created successfully!"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.HealthResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "💪 Shortener API is up and running!"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handler.LinkSample": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer",
                    "example": 10
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-02T00:00:00Z"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2027-04-08T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "long_url": {
                    "type": "string",
                    "example": "https://facebook.com"
                },
                "short_code": {
                    "type": "string",
                    "example": "1234567890"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-09-11T00:00:00Z"
                }
            }
        },
        "handler.OptionalExpiresAt": {
            "type": "object",
            "properties": {
                "present": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateLinkRequest": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "$ref": "#/definitions/handler.OptionalExpiresAt"
                },
                "long_url": {
                    "type": "string",
                    "maxLength": 2048
                }
            }
        },
        "handler.UpdateLinkResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.LinkSample"
                },
                "message": {
                    "type": "string",
                    "example": "Link updated successfully!"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}